+++ src/main/webapp/WEB-INF/views/inventeur/list.jsp
@@ -5,6 +5,12 @@
 		<a href="ReprotsController?filename=RptInventeur" class="btn btn-light mb-1 ml-1"><i class="fa fa-print mr-1"></i>
 			Imprimer</a>
 	</div>
+	<c:if test="${status == 'deleted'}">
+		<div class="alert alert-success" role="alert">Inventeur supprime avec succes.</div>
+	</c:if>
+	<c:if test="${not empty globalError}">
+		<div class="alert alert-danger" role="alert">${globalError}</div>
+	</c:if>
 	<div class="table-responsive">
 		<table class="table">
 			<thead class="thead-dark">
@@ -19,6 +25,11 @@
 				</tr>
 			</thead>
 			<tbody>
+				<c:if test="${empty inventeurs}">
+					<tr>
+						<td colspan="7" class="text-center text-muted py-4">Aucun inventeur disponible.</td>
+					</tr>
+				</c:if>
 				<c:forEach items="${inventeurs}" var="inventeur">
 					<tr>
 						<td scope="col">${inventeur.nom} ${inventeur.prenom}</td>
@@ -28,7 +39,7 @@
 						<td scope="col">${inventeur.entreprise.nom}</td>
 						<td><a class="text-success" href="?mode=updating&id=${inventeur.num}&page=${currentPage}"><i class="fa fa-edit"></i></a>
 						</td>
-						<td><a href="?mode=delete&id=${inventeur.num}&page=${currentPage}"><i class="fa fa-trash"></i></a></td>
+						<td><a href="?mode=delete&id=${inventeur.num}&page=${currentPage}" data-confirm-delete="true" data-delete-label="cet inventeur"><i class="fa fa-trash"></i></a></td>
 					</tr>
 				</c:forEach>
 			</tbody>
@@ -39,15 +50,15 @@
 		<nav aria-label="Pagination des inventeurs" class="mt-2">
 			<ul class="pagination justify-content-center">
 				<li class="page-item ${currentPage == 1 ? 'disabled' : ''}">
-					<a class="page-link" href="?mode=list&page=${currentPage - 1}">Precedent</a>
+					<a class="page-link" href="?mode=list&page=${currentPage - 1}" data-loading-link="true">Precedent</a>
 				</li>
 				<c:forEach begin="1" end="${totalPages}" var="p">
 					<li class="page-item ${currentPage == p ? 'active' : ''}">
-						<a class="page-link" href="?mode=list&page=${p}">${p}</a>
+						<a class="page-link" href="?mode=list&page=${p}" data-loading-link="true">${p}</a>
 					</li>
 				</c:forEach>
 				<li class="page-item ${currentPage == totalPages ? 'disabled' : ''}">
-					<a class="page-link" href="?mode=list&page=${currentPage + 1}">Suivant</a>
+					<a class="page-link" href="?mode=list&page=${currentPage + 1}" data-loading-link="true">Suivant</a>
 				</li>
 			</ul>
 		</nav>
